Generalize a cache for downloading files
Related to #620
